Unit accessibility data in UnitSerializer.
